Update: Fix rendering issues
