+++ src/components/admin/combo/ComboModal.js
@@ -1,15 +1,14 @@
-import { Modal, ModalHeader, ModalBody, ModalFooter, Button, TextInput, Select } from 'flowbite-react';
+import { Modal, ModalHeader, ModalBody, ModalFooter, Button, TextInput } from 'flowbite-react';
 import { useState, useEffect, useMemo } from 'react';
 
-const ComboModal = ({ isOpen, onClose, onAddCombo, onUpdateCombo, comboToEdit, products = [] }) => {
+const ComboModal = ({ isOpen, onClose, onAddCombo, comboToEdit, products = [] }) => {
     const [combo, setCombo] = useState({
         name: '',
         description: '',
         price: '',
-        stock: '',
-        productCombo: [] // Lista para productos y cantidades
+        productCombo: [],
     });
-
+    
     const [file, setFile] = useState(null);
     const [query, setQuery] = useState('');
 
@@ -19,14 +18,13 @@ const ComboModal = ({ isOpen, onClose, onAddCombo, onUpdateCombo, comboToEdit, p
                 name: comboToEdit.name,
                 description: comboToEdit.description,
                 price: comboToEdit.price,
-                stock: comboToEdit.stock,
-                productCombo: comboToEdit.productCombo || [] // Asignar productos del combo a editar
+                productCombo: comboToEdit.productCombo || []
             });
         } else {
-            setCombo({ name: '', description: '', price: '', stock: '', productCombo: [] });
+            setCombo({ idCombo: '', name: '', description: '', price: '', productCombo: [] });
             setFile(null);
         }
-    }, [comboToEdit, products]);
+    }, [comboToEdit]);
 
     useEffect(() => {
         if (!isOpen) {
@@ -54,50 +52,16 @@ const ComboModal = ({ isOpen, onClose, onAddCombo, onUpdateCombo, comboToEdit, p
         setFile(selectedFile);
     };
 
-    const handleProductSelect = (productId) => {
-        const productExists = combo.productCombo.find(item => item.idProduct === productId);
-        if (!productExists) {
-            setCombo(prevCombo => ({
-                ...prevCombo,
-                productCombo: [...prevCombo.productCombo, { idProduct: productId, quantity: 1 }] // Añadir nuevo producto con cantidad inicial 1
-            }));
-        }
-    };
-
-    const handleQuantityChange = (productId, quantity) => {
-        setCombo(prevCombo => ({
-            ...prevCombo,
-            productCombo: prevCombo.productCombo.map(item => 
-                item.idProduct === productId ? { ...item, quantity: Number(quantity) } : item
-            )
-        }));
-    };
-
-    const handleAddProduct = () => {
-        if (combo.productCombo.length < 5) { // Limitar a un máximo de 5 productos, puedes cambiarlo según tu necesidad
-            setCombo(prevCombo => ({
-                ...prevCombo,
-                productCombo: [...prevCombo.productCombo, { idProduct: null, quantity: 1 }] // Añadir nuevo producto vacío
-            }));
-        }
-    };
-
-    const handleSubmit = async (e) => {
+    const handleSubmitCombo = async (e) => {
         e.preventDefault();
         try {
             const comboJson = {
                 name: combo.name,
                 description: combo.description,
                 price: combo.price,
-                stock: combo.stock,
-                productCombo: combo.productCombo // Incluir lista de productos y cantidades
+                productCombo: combo.productCombo
             };
-
-            if (comboToEdit) {
-                await onUpdateCombo(comboToEdit.idCombo, comboJson, file);
-            } else {
-                await onAddCombo(comboJson, file);
-            }
+            await onAddCombo(comboJson, file);
             onClose();
         } catch (error) {
             console.error('Error al agregar/actualizar combo:', error);
@@ -110,7 +74,7 @@ const ComboModal = ({ isOpen, onClose, onAddCombo, onUpdateCombo, comboToEdit, p
                 <p className='poppins-bold text-2xl'>{comboToEdit ? 'Editar Combo' : 'Agregar Combo'}</p>
             </ModalHeader>
             <ModalBody>
-                <form onSubmit={handleSubmit}>
+                <form onSubmit={handleSubmitCombo}>
                     <div className="mb-4">
                         <label htmlFor="name" className="block text-sm font-medium text-gray-700">Nombre</label>
                         <TextInput
@@ -144,46 +108,6 @@ const ComboModal = ({ isOpen, onClose, onAddCombo, onUpdateCombo, comboToEdit, p
                             placeholder="Ingrese el precio del combo."
                             required
                         />
-                    </div>
-                    <div className="mb-4">
-                        <label htmlFor="stock" className="block text-sm font-medium text-gray-700">Stock</label>
-                        <TextInput
-                            id="stock"
-                            name="stock"
-                            type="number"
-                            value={combo.stock}
-                            onChange={handleInputChange}
-                            placeholder="Ingrese el stock del combo."
-                            required
-                        />
-                    </div>
-                    <div className="mb-4">
-                        <label htmlFor="dropdown" className="block text-sm font-medium text-gray-700">Seleccionar Productos</label>
-                        {combo.productCombo.map((item, index) => (
-                            <div key={index} className="flex items-center mb-2">
-                                <Select
-                                    onChange={(e) => handleProductSelect(e.target.value)}
-                                    className="mr-2"
-                                    value={item.idProduct || ''}
-                                >
-                                    <option disabled value="">Seleccionar producto</option>
-                                    {filteredOptions.map(product => (
-                                        <option key={product.idProduct} value={product.idProduct}>
-                                            {product.name}
-                                        </option>
-                                    ))}
-                                </Select>
-                                <TextInput
-                                    type="number"
-                                    value={item.quantity}
-                                    onChange={(e) => handleQuantityChange(item.idProduct, e.target.value)}
-                                    placeholder="Cantidad"
-                                    className="w-20"
-                                    required
-                                />
-                            </div>
-                        ))}
-                        <Button onClick={handleAddProduct} className="rounded-full p-2">+</Button>
                     </div>
                     <div className="mb-4">
                         <label htmlFor="file" className="block text-sm font-medium text-gray-700">Imagen</label>
