+++ src/containers/admin/category/CategoryContainer.js
@@ -1,77 +1,55 @@
 import { useEffect, useState } from "react";
+import CategoryList from "../../../components/admin/category/CategoryList";
+import AddCategoryModal from "../../../components/admin/category/AddCategoryModal";
+import { Spinner } from 'flowbite-react'; 
 import { IoMdAdd } from "react-icons/io";
-import ErrorModal from "../ErrorModal";
+import ErrorModal from "../../../components/admin/ErrorModal";
 import { FaSearch } from "react-icons/fa";
-import { Spinner } from "flowbite-react";
-import ProductList from "./ProductList"
-import ProductModal from "./ProductModal"
 
-const ProductContainer = () => {
-    const [subcategories, setSubcategories] = useState([]);
-    const [products, setProducts] = useState([]);
-    const [loading, setLoading] = useState(true);
+
+const CategoryContainer = () => {
+    const [categories, setCategories] = useState([]);
     const [error, setError] = useState(null);
-    const [isErrorModalOpen, setIsErrorModalOpen] = useState(false); 
+    const [loading, setLoading] = useState(true);
     const [isModalOpen, setIsModalOpen] = useState(false);
-    const [editingProduct, setEditingProduct] = useState(null);
+    const [editingCategory, setEditingCategory] = useState(null);
+    const [isErrorModalOpen, setIsErrorModalOpen] = useState(false); 
     const [searchTerm, setSearchTerm] = useState("");
 
-    useEffect(() => {
-        const fetchSubcategories = async () => {
-            try {
-                const response = await fetch("http://localhost:8080/api-namp/subcategory");
-                if (!response.ok) {
-                    throw new Error('Error al obtener las subcategorías');
-                }
-                const data = await response.json();
-                setSubcategories(data);
-            } catch (error) {
-                setError(error.message);
-                setIsErrorModalOpen(true);
-            }
-        };
 
-        fetchProduct();
-        fetchSubcategories();
+    useEffect(() => {
+        fetchCategories();
     }, []);
-    
-    const fetchProduct = async () => {
+
+    const fetchCategories = async () => {
         setLoading(true);
         try {
-            const response = await fetch("http://localhost:8080/api-namp/product", {
-                method: 'GET',
-                mode: 'cors' 
-            });
-
+            const response = await fetch("http://localhost:8080/api-namp/category");
             if (!response.ok) {
-                throw new Error('Error al obtener las subcategorías');
+                throw new Error('Error al traer las categorías');
             }
-
             const data = await response.json();
-
-            setProducts(data);
+            setCategories(data);
         } catch (error) {
             setError(error.message);
             setIsErrorModalOpen(true);
         } finally {
-            setTimeout(() => {
+            setTimeout(()=>{
                 setLoading(false);
-            }, 800); 
+            },800)
         }
     };
 
-    const addProduct = async (product, file) => {
+    const addCategory = async (newCategory) => {
         setLoading(true);
         try {
-            const formData = new FormData();
-            formData.append('product', JSON.stringify(product));
-            formData.append('file', file);
-
-            const response = await fetch("http://localhost:8080/api-namp/product", {
+            const response = await fetch("http://localhost:8080/api-namp/category", {
                 method: "POST",
-                body: formData
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify(newCategory)
             });
-
             if (!response.ok) {
                 const errorText = await response.text();
 
@@ -83,11 +61,9 @@ const ProductContainer = () => {
                     errorMessage = errorText; // Es un mensaje de error simple
                 }
 
-                throw new Error(errorMessage || 'Error al agregar el producto');
+                throw new Error(errorMessage || 'Error al agregar la categoría');
             }
-            
-            await fetchProduct();
-            
+            await fetchCategories();
             setIsModalOpen(false);
         } catch (error) {
             setError(error.message);
@@ -97,20 +73,16 @@ const ProductContainer = () => {
         }
     };
 
-    const updateProduct = async (id, updatedProduct, file) => {
+    const updateCategory = async (id, updateCategory) => {
         setLoading(true);
         try {
-            const formData = new FormData();
-            formData.append('product', JSON.stringify(updatedProduct));
-            if (file) {
-                formData.append('file', file);
-            }
-    
-            const response = await fetch(`http://localhost:8080/api-namp/product/${id}`, {
+            const response = await fetch(`http://localhost:8080/api-namp/category/${id}`, {
                 method: "PUT",
-                body: formData
+                headers: {
+                    "Content-Type": "application/json"
+                },
+                body: JSON.stringify(updateCategory)
             });
-
             if (!response.ok) {
                 const errorText = await response.text();
 
@@ -121,11 +93,10 @@ const ProductContainer = () => {
                 } else {
                     errorMessage = errorText; // Es un mensaje de error simple
                 }
-                
-                throw new Error(errorMessage || 'Error al actualizar el producto');
-            }
 
-            await fetchProduct();
+                throw new Error(errorMessage || 'Error al actualizar la categoría');
+            }
+            await fetchCategories();
             setIsModalOpen(false); 
         } catch (error) {
             setError(error.message);
@@ -134,18 +105,17 @@ const ProductContainer = () => {
             setLoading(false);
         }
     };
-    
-    const deleteProduct = async (id) => {
+
+    const deleteCategory = async (id) => {
         setLoading(true);
         try {
-            
-            const response = await fetch(`http://localhost:8080/api-namp/product/${id}`, {
+            const response = await fetch(`http://localhost:8080/api-namp/category/${id}`, {
                 method: "DELETE"
             });
             if (!response.ok) {
-                throw new Error('Error al eliminar el producto');
+                throw new Error('Error al eliminar la categoría');
             }
-            await fetchProduct();
+            await fetchCategories();
         } catch (error) {
             setError(error.message);
             setIsErrorModalOpen(true); 
@@ -161,16 +131,14 @@ const ProductContainer = () => {
         const match = errorText.match(regex);
         return match ? match[1] : null;
     };
-    
 
-
-    const handleAddProductClick = () => {
-        setEditingProduct(null);
+    const handleAddCategoryClick = () => {
+        setEditingCategory(null);
         setIsModalOpen(true);
     };
 
-    const editProductHandler = (subcategory) => {
-        setEditingProduct(subcategory);
+    const editCategoryHandler = (category) => {
+        setEditingCategory(category);
         setIsModalOpen(true);
     };
 
@@ -179,6 +147,10 @@ const ProductContainer = () => {
         setError(null);
     };
 
+    const filteredCategories = categories.filter(category =>
+        category.name.toLowerCase().includes(searchTerm.toLowerCase())
+    );
+
     if (loading) {
         return (
             <div className="bottom-1/2 flex justify-center items-center h-80">
@@ -186,7 +158,8 @@ const ProductContainer = () => {
             </div>
         )
     }
-    return ( 
+
+    return (
         <div className="mb-4">
             <div className="flex justify-between mr-4 mt-4 gap-2">
                 <div className="flex items-center gap-2">
@@ -201,35 +174,33 @@ const ProductContainer = () => {
                 </div>
                 
                 <button
-                    onClick={handleAddProductClick}
+                    onClick={handleAddCategoryClick}
                     className="bg-blue-600 hover:bg-blue-500 text-white font-semibold rounded-lg p-2 flex items-center"
                 >
                     <span><IoMdAdd/></span>
                     Agregar
                 </button>
             </div>
-            <ProductList
-                products={products}
-                updateProduct={updateProduct}
-                deleteProduct={deleteProduct}
-                addProduct={addProduct}
-                onEditProduct={editProductHandler}
+            <CategoryList
+                categories={filteredCategories}
+                updateCategory={updateCategory}
+                deleteCategory={deleteCategory}
+                addCategory={addCategory}
+                onEditCategory={editCategoryHandler}
             />
-            <ProductModal
+            <AddCategoryModal
                 isOpen={isModalOpen}
                 onClose={() => setIsModalOpen(false)}
-                onAddProduct={addProduct}
-                onUpdateProduct={updateProduct}
-                productToEdit={editingProduct}
-                subcategories={subcategories}
+                onAddCategory={addCategory}
+                onUpdateCategory={updateCategory}
+                categoryToEdit={editingCategory}
             />
             <ErrorModal 
                 isErrorModalOpen={isErrorModalOpen} closeErrorModal={closeErrorModal} 
                 error={error}
             />
-            
         </div>
     );
-}
- 
-export default ProductContainer;
+};
+
+export default CategoryContainer;
