+++ src/containers/admin/product/ProductContainer.js
@@ -1,55 +1,77 @@
 import { useEffect, useState } from "react";
-import CategoryList from "./CategoryList";
-import AddCategoryModal from "./AddCategoryModal";
-import { Spinner } from 'flowbite-react'; 
 import { IoMdAdd } from "react-icons/io";
-import ErrorModal from "../ErrorModal";
+import ErrorModal from "../../../components/admin/ErrorModal";
 import { FaSearch } from "react-icons/fa";
+import { Spinner } from "flowbite-react";
+import ProductList from "../../../components/admin/product/ProductList"
+import ProductModal from "../../../components/admin/product/ProductModal"
 
-
-const CategoryContainer = () => {
-    const [categories, setCategories] = useState([]);
-    const [error, setError] = useState(null);
+const ProductContainer = () => {
+    const [subcategories, setSubcategories] = useState([]);
+    const [products, setProducts] = useState([]);
     const [loading, setLoading] = useState(true);
-    const [isModalOpen, setIsModalOpen] = useState(false);
-    const [editingCategory, setEditingCategory] = useState(null);
+    const [error, setError] = useState(null);
     const [isErrorModalOpen, setIsErrorModalOpen] = useState(false); 
+    const [isModalOpen, setIsModalOpen] = useState(false);
+    const [editingProduct, setEditingProduct] = useState(null);
     const [searchTerm, setSearchTerm] = useState("");
 
-
     useEffect(() => {
-        fetchCategories();
-    }, []);
+        const fetchSubcategories = async () => {
+            try {
+                const response = await fetch("http://localhost:8080/api-namp/subcategory");
+                if (!response.ok) {
+                    throw new Error('Error al obtener las subcategorías');
+                }
+                const data = await response.json();
+                setSubcategories(data);
+            } catch (error) {
+                setError(error.message);
+                setIsErrorModalOpen(true);
+            }
+        };
 
-    const fetchCategories = async () => {
+        fetchProduct();
+        fetchSubcategories();
+    }, []);
+    
+    const fetchProduct = async () => {
         setLoading(true);
         try {
-            const response = await fetch("http://localhost:8080/api-namp/category");
+            const response = await fetch("http://localhost:8080/api-namp/product", {
+                method: 'GET',
+                mode: 'cors' 
+            });
+
             if (!response.ok) {
-                throw new Error('Error al traer las categorías');
+                throw new Error('Error al obtener las subcategorías');
             }
+
             const data = await response.json();
-            setCategories(data);
+
+            setProducts(data);
         } catch (error) {
             setError(error.message);
             setIsErrorModalOpen(true);
         } finally {
-            setTimeout(()=>{
+            setTimeout(() => {
                 setLoading(false);
-            },800)
+            }, 800); 
         }
     };
 
-    const addCategory = async (newCategory) => {
+    const addProduct = async (product, file) => {
         setLoading(true);
         try {
-            const response = await fetch("http://localhost:8080/api-namp/category", {
+            const formData = new FormData();
+            formData.append('product', JSON.stringify(product));
+            formData.append('file', file);
+
+            const response = await fetch("http://localhost:8080/api-namp/product", {
                 method: "POST",
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify(newCategory)
+                body: formData
             });
+
             if (!response.ok) {
                 const errorText = await response.text();
 
@@ -61,9 +83,11 @@ const CategoryContainer = () => {
                     errorMessage = errorText; // Es un mensaje de error simple
                 }
 
-                throw new Error(errorMessage || 'Error al agregar la categoría');
+                throw new Error(errorMessage || 'Error al agregar el producto');
             }
-            await fetchCategories();
+            
+            await fetchProduct();
+            
             setIsModalOpen(false);
         } catch (error) {
             setError(error.message);
@@ -73,16 +97,20 @@ const CategoryContainer = () => {
         }
     };
 
-    const updateCategory = async (id, updateCategory) => {
+    const updateProduct = async (id, updatedProduct, file) => {
         setLoading(true);
         try {
-            const response = await fetch(`http://localhost:8080/api-namp/category/${id}`, {
+            const formData = new FormData();
+            formData.append('product', JSON.stringify(updatedProduct));
+            if (file) {
+                formData.append('file', file);
+            }
+    
+            const response = await fetch(`http://localhost:8080/api-namp/product/${id}`, {
                 method: "PUT",
-                headers: {
-                    "Content-Type": "application/json"
-                },
-                body: JSON.stringify(updateCategory)
+                body: formData
             });
+
             if (!response.ok) {
                 const errorText = await response.text();
 
@@ -93,10 +121,11 @@ const CategoryContainer = () => {
                 } else {
                     errorMessage = errorText; // Es un mensaje de error simple
                 }
-
-                throw new Error(errorMessage || 'Error al actualizar la categoría');
+                
+                throw new Error(errorMessage || 'Error al actualizar el producto');
             }
-            await fetchCategories();
+
+            await fetchProduct();
             setIsModalOpen(false); 
         } catch (error) {
             setError(error.message);
@@ -105,17 +134,18 @@ const CategoryContainer = () => {
             setLoading(false);
         }
     };
-
-    const deleteCategory = async (id) => {
+    
+    const deleteProduct = async (id) => {
         setLoading(true);
         try {
-            const response = await fetch(`http://localhost:8080/api-namp/category/${id}`, {
+            
+            const response = await fetch(`http://localhost:8080/api-namp/product/${id}`, {
                 method: "DELETE"
             });
             if (!response.ok) {
-                throw new Error('Error al eliminar la categoría');
+                throw new Error('Error al eliminar el producto');
             }
-            await fetchCategories();
+            await fetchProduct();
         } catch (error) {
             setError(error.message);
             setIsErrorModalOpen(true); 
@@ -131,14 +161,16 @@ const CategoryContainer = () => {
         const match = errorText.match(regex);
         return match ? match[1] : null;
     };
+    
 
-    const handleAddCategoryClick = () => {
-        setEditingCategory(null);
+
+    const handleAddProductClick = () => {
+        setEditingProduct(null);
         setIsModalOpen(true);
     };
 
-    const editCategoryHandler = (category) => {
-        setEditingCategory(category);
+    const editProductHandler = (subcategory) => {
+        setEditingProduct(subcategory);
         setIsModalOpen(true);
     };
 
@@ -147,10 +179,6 @@ const CategoryContainer = () => {
         setError(null);
     };
 
-    const filteredCategories = categories.filter(category =>
-        category.name.toLowerCase().includes(searchTerm.toLowerCase())
-    );
-
     if (loading) {
         return (
             <div className="bottom-1/2 flex justify-center items-center h-80">
@@ -158,8 +186,7 @@ const CategoryContainer = () => {
             </div>
         )
     }
-
-    return (
+    return ( 
         <div className="mb-4">
             <div className="flex justify-between mr-4 mt-4 gap-2">
                 <div className="flex items-center gap-2">
@@ -174,33 +201,35 @@ const CategoryContainer = () => {
                 </div>
                 
                 <button
-                    onClick={handleAddCategoryClick}
+                    onClick={handleAddProductClick}
                     className="bg-blue-600 hover:bg-blue-500 text-white font-semibold rounded-lg p-2 flex items-center"
                 >
                     <span><IoMdAdd/></span>
                     Agregar
                 </button>
             </div>
-            <CategoryList
-                categories={filteredCategories}
-                updateCategory={updateCategory}
-                deleteCategory={deleteCategory}
-                addCategory={addCategory}
-                onEditCategory={editCategoryHandler}
+            <ProductList
+                products={products}
+                updateProduct={updateProduct}
+                deleteProduct={deleteProduct}
+                addProduct={addProduct}
+                onEditProduct={editProductHandler}
             />
-            <AddCategoryModal
+            <ProductModal
                 isOpen={isModalOpen}
                 onClose={() => setIsModalOpen(false)}
-                onAddCategory={addCategory}
-                onUpdateCategory={updateCategory}
-                categoryToEdit={editingCategory}
+                onAddProduct={addProduct}
+                onUpdateProduct={updateProduct}
+                productToEdit={editingProduct}
+                subcategories={subcategories}
             />
             <ErrorModal 
                 isErrorModalOpen={isErrorModalOpen} closeErrorModal={closeErrorModal} 
                 error={error}
             />
+            
         </div>
     );
-};
-
-export default CategoryContainer;
+}
+ 
+export default ProductContainer;
