+++ src/pages/private/CategoryPage.js
@@ -1,6 +1,6 @@
 import { FaThList } from "react-icons/fa";
 
-import CategoryContainer from "../../components/admin/category/CategoryContainer";
+import CategoryContainer from "../../containers/admin/category/CategoryContainer";
 
 const CategoryPage = () => {
     return (
